+++ test/lib/javascript/get-prop-types-identifier-name.js
@@ -1,0 +1,24 @@
+const { parse } = require('@babel/parser');
+const test = require('ava');
+
+const getPropTypesIdentifierName = require('../../../lib/parse/javascript/get-prop-types-identifier-name');
+
+test('With PropTypes import', t => {
+  const syntaxTree = parse('import pt from "prop-types";', {
+    sourceType: 'module'
+  });
+  t.is(
+    getPropTypesIdentifierName({ syntaxTree }).propTypesIdentifierName,
+    'pt'
+  );
+});
+
+test('Without PropTypes import', t => {
+  const syntaxTree = parse('import React from "react";', {
+    sourceType: 'module'
+  });
+  t.is(
+    getPropTypesIdentifierName({ syntaxTree }).propTypesIdentifierName,
+    undefined
+  );
+});