+++ webpack-plugin/generate-classes.js
@@ -2,17 +2,14 @@ const fs = require('fs');
 
 const generateClass = require('../lib');
 
-const attemptGenerateClass = ({ baseClass, modulePath, namespace, indent }) => {
+const attemptGenerateClass = (modulePath, options) => {
   try {
     const sourceCode = fs.readFileSync(modulePath, 'utf-8');
-    const { code, componentName } = generateClass({
-      baseClass,
-      indent,
-      namespace,
-      sourceCode
-    });
+    const { code, className } = generateClass(
+      Object.assign({}, options, { sourceCode })
+    );
 
-    return { code, componentName };
+    return { code, className };
   } catch (error) {
     return {
       error: `\n${modulePath}\n${error.message}\n`
@@ -20,21 +17,21 @@ const attemptGenerateClass = ({ baseClass, modulePath, namespace, indent }) => {
   }
 };
 
-const generateClasses = ({ baseClass, indent, modulePaths, namespace }) => {
+const generateClasses = ({ modulePaths, options }) => {
   const startTime = new Date().getTime();
   const classes = modulePaths.map(modulePath =>
-    attemptGenerateClass({ baseClass, indent, modulePath, namespace })
+    attemptGenerateClass(modulePath, options)
   );
-  const duplicates = classes.reduce((accum, { componentName }, index) => {
-    if (componentName) {
+  const duplicates = classes.reduce((accum, { className }, index) => {
+    if (className) {
       const duplicateIndex = classes
         .slice(index + 1)
-        .findIndex(c => c.componentName === componentName);
+        .findIndex(c => c.className === className);
 
       if (duplicateIndex !== -1) {
         return accum.concat(
-          `${componentName} (${modulePaths[index]})`,
-          `${componentName} (${modulePaths[duplicateIndex + 1]})`
+          `${className} (${modulePaths[index]})`,
+          `${className} (${modulePaths[duplicateIndex + 1]})`
         );
       }
     }
@@ -54,10 +51,10 @@ const generateClasses = ({ baseClass, indent, modulePaths, namespace }) => {
   };
 };
 
-// Hook for running in parallel with child_process. Expects the same options object as 'generateClasses' above.
-process.on('message', options => {
-  if (options.modulePaths) {
-    process.send(generateClasses(options));
+// Hook for running in parallel with child_process. Expects the same arguments as 'generateClasses' above.
+process.on('message', ({ modulePaths, options }) => {
+  if (modulePaths) {
+    process.send(generateClasses({ modulePaths, options }));
   }
 });