+++ webpack-plugin/index.js
@@ -68,22 +68,22 @@ PropTypesCSharpPlugin.prototype.apply = function(compiler) {
       this.options.exclude
     );
 
-    const generateClassesOptions = {
+    const options = {
       baseClass: this.options.baseClass,
       indent: this.options.indent,
-      modulePaths,
-      namespace: this.options.namespace
+      namespace: this.options.namespace,
+      parser: this.options.parser
     };
 
     if (!isAsync) {
-      const result = generateClasses(generateClassesOptions);
+      const result = generateClasses({ modulePaths, options });
       const { classes, error } = result;
 
       if (!error) {
-        classes.forEach(({ code, componentName }) => {
-          if (code && componentName) {
+        classes.forEach(({ code, className }) => {
+          if (code && className) {
             compilation.assets[
-              path.join(this.outputPath, `${componentName}.cs`)
+              path.join(this.outputPath, `${className}.cs`)
             ] = {
               source: () => code,
               size: () => code.length
@@ -95,7 +95,7 @@ PropTypesCSharpPlugin.prototype.apply = function(compiler) {
       log(this.options, isAsync, compilation, result);
     } else {
       // Run class generation in parallel
-      generateClassesAsync.send(generateClassesOptions);
+      generateClassesAsync.send({ modulePaths, options });
     }
   };
 
@@ -110,11 +110,11 @@ PropTypesCSharpPlugin.prototype.apply = function(compiler) {
 
       // Write files to disk since webpack dev server doesn't do so
       if (!error) {
-        classes.forEach(({ code, componentName }) => {
-          if (code && componentName) {
+        classes.forEach(({ code, className }) => {
+          if (code && className) {
             const basePath = path.join(compiler.outputPath, this.outputPath);
             fsExtra.ensureDirSync(basePath);
-            fs.writeFileSync(path.join(basePath, `${componentName}.cs`), code);
+            fs.writeFileSync(path.join(basePath, `${className}.cs`), code);
           }
         });
       }